+++ app/Http/Requests/EmpresaRequest.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class EmpresaRequest extends FormRequest
+{
+    /**
+    * Determine if the user is authorized to make this request.
+    *
+    * @return bool
+    */
+    public function authorize()
+    {
+        return true;
+    }
+    
+    /**
+    * Get the validation rules that apply to the request.
+    *
+    * @return array<string, mixed>
+    */
+    public function rules()
+    {
+        return [
+            'nome'      =>  ['required'],
+            'descricao' =>  ['required'],
+            'logo'      =>  ['required'],
+            'imagem'    =>  ['required'],
+            'telefone'  =>  ['required']
+        ];
+    }
+}
